pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ijepath"
version = "0.1.0"
description = "Pathology-focused adaptation of I-JEPA for cross-resolution self-supervised learning on whole-slide images."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "I-JEPATH contributors" }
]
dependencies = [
  "numpy",
  "pandas",
  "Pillow",
  "PyYAML",
  "scikit-learn",
]

[project.optional-dependencies]
train = [
  "opencv-python",
  "submitit",
  "tqdm",
  "torch",
  "torchvision",
  "wandb",
  "wholeslidedata",
]
dev = [
  "build",
  "pytest",
  "twine",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["configs", "configs.*", "ijepath", "ijepath.*"]

[tool.setuptools.package-data]
configs = ["*.yaml", "profiles/*.yaml", "runs/*.yaml"]
